Refactor interface PrivilegeLoader
